TST: Fix wrong type on C++ side

+++ tests/src/test_arr_to_mat.h
@@ -8,7 +8,7 @@ namespace carma {
 namespace tests {
 
 int test_arr_to_mat_double_copy(py::array_t<double> arr);
-int test_arr_to_mat_1d(py::array_t<long>& arr, bool copy, bool strict);
+int test_arr_to_mat_1d(py::array_t<double>& arr, bool copy, bool strict);
 int test_arr_to_mat_long(py::array_t<long>& arr, bool copy, bool strict);
 int test_arr_to_mat_double(py::array_t<double>& arr, bool copy, bool strict);
 int test_arr_to_col(py::array_t<double>& arr, bool copy, bool strict);
